TemplateApp/pyproject.toml
[project]
name = "templateapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyserial",
    "PyQt6",
    "PyQt6-Charts"
]
